Add CohortsEntity trait
Apply it to Course and User.
